# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SchoolCore - Multi-tenant school management platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /schools/{schoolId}/students:
    get:
      operationId: listStudents
      tags: [students]
      summary: List students in a school
      parameters:
        - name: schoolId
          in: path
          required: true
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
        - name: classId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
            enum: [active, inactive]
      responses:
        "200":
          description: List of students
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Student"
    post:
      operationId: createStudent
      tags: [students]
      summary: Add a new student
      parameters:
        - name: schoolId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudentInput"
      responses:
        "201":
          description: Student created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
  /schools/{schoolId}/students/{studentId}:
    get:
      operationId: getStudent
      tags: [students]
      summary: Get student details
      parameters:
        - name: schoolId
          in: path
          required: true
          schema:
            type: integer
        - name: studentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Student detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentDetail"
    patch:
      operationId: updateStudent
      tags: [students]
      summary: Update student info
      parameters:
        - name: schoolId
          in: path
          required: true
          schema:
            type: integer
        - name: studentId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudentUpdate"
      responses:
        "200":
          description: Updated student
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
    delete:
      operationId: deleteStudent
      tags: [students]
      summary: Remove a student
      parameters:
        - name: schoolId
          in: path
          required: true
          schema:
            type: integer
        - name: studentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                required: [success]

  # ─── TEACHERS ────────────────────────────────────────────────────────────────
  /schools/{schoolId}/teachers:
    get:
      operationId: listTeachers
      tags: [teachers]
      summary: List teachers in a school
      parameters:
        - name: schoolId
          in: path
          required: true
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of teachers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Teacher"
    post:
      operationId: createTeacher
      tags: [teachers]
      summary: Add a teacher
      parameters:
        - name: schoolId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TeacherInput"
      responses:
        "201":
          description: Teacher created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Teacher"
  /schools/{schoolId}/teachers/{teacherId}:
    get:
      operationId: getTeacher
      tags: [teachers]
      summary: Get teacher details
      parameters:
        - name: schoolId
          in: path
          required: true
          schema:
            type: integer
        - name: teacherId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Teacher detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Teacher"
    patch:
      operationId: updateTeacher
      tags: [teachers]
      summary: Update teacher
      parameters:
        - name: schoolId
          in: path
          required: true
          schema:
            type: integer
        - name: teacherId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TeacherUpdate"
      responses:
        "200":
          description: Updated teacher
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Teacher"

  # ─── CLASSES ─────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Student:
      type: object
      required: [id, firstName, lastName, schoolId, status, createdAt]
      properties:
        id:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        dateOfBirth:
          type: ["string", "null"]
        photoUrl:
          type: ["string", "null"]
        classId:
          type: ["integer", "null"]
        className:
          type: ["string", "null"]
        parentPhone:
          type: ["string", "null"]
        parentRelationship:
          type: ["string", "null"]
        schoolId:
          type: integer
        status:
          type: string
          enum: [active, inactive]
        createdAt:
          type: string
          format: date-time
...
    StudentDetail:
      allOf:
        - $ref: "#/components/schemas/Student"
        - type: object
          properties:
            attendanceRate:
              type: number
            averageScore:
              type: number
            feesBalance:
              type: number
    StudentInput:
      type: object
      required: [firstName, lastName]
      properties:
        firstName:
          type: string
        lastName:
          type: string
        dateOfBirth:
          type: string
          format: date
        classId:
          type: integer
        parentPhone:
          type: string
        parentRelationship:
          type: string
        photoUrl:
          type: ["string", "null"]
    StudentUpdate:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        dateOfBirth:
          type: string
          format: date
        classId:
          type: integer
        parentPhone:
          type: string
        parentRelationship:
          type: string
        status:
          type: string
          enum: [active, inactive]
        photoUrl:
          type: ["string", "null"]

    # ─── TEACHERS ──────────────────────────────────────────────────────────────
    Teacher:
      type: object
      required: [id, firstName, lastName, schoolId, phone, createdAt]
      properties:
        id:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
        email:
          type: ["string", "null"]
        subjects:
          type: array
          items:
            type: string
        classes:
          type: array
          items:
            type: string
        schoolId:
          type: integer
        status:
          type: string
          enum: [active, inactive]
        createdAt:
          type: string
          format: date-time
    TeacherInput:
      type: object
      required: [firstName, lastName, phone]
      properties:
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
        email:
          type: ["string", "null"]
    TeacherUpdate:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
        email:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, inactive]

    # ─── CLASSES ───────────────────────────────────────────────────────────────
...